requirements.txt
httpx
pydantic
PyJWT

